Share x509 parsing between ring and rc_crypto
