+++ rs-client-demo/src/main.rs
@@ -1,7 +1,6 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-use env_logger;
 use remote_settings_client::{Client, Collection, SignatureError, Verification};
 use serde::Deserialize;
 pub use url::{ParseError, Url};
